add useLayoutSeparators hook to utilize separator logic
